+++ src/bluesky/application/dto/mod.rs
@@ -3,3 +3,4 @@ pub mod commit_message_dto;
 pub mod embed_dto;
 pub mod facet_dto;
 pub mod login_credential_dto;
+pub mod post_actions_dto;