+++ src/util.rs
@@ -1,4 +1,5 @@
 use crate::Message;
+use crate::bluesky::domain::message::model::post_message::PostMessage;
 use chrono::Utc;
 use curl::easy::List;
 use regex::{Captures, Regex};
@@ -31,11 +32,25 @@ pub fn set_headers(header_list: Vec<String>) -> List {
     headers
 }
 
-pub fn read_json_file(file_path: &str) -> Result<Message, Error> {
+pub fn message_from_json_file(file_path: &str) -> Result<Message, Error> {
     let file = File::open(file_path).expect("File not found");
     let reader = BufReader::new(file);
 
     let json_object: Message = serde_json::from_reader(reader)?;
 
     Ok(json_object)
+}
+
+pub fn merge_message(message_from_json: &Message, message_from_args: &PostMessage) -> Message {
+    let message = if !message_from_args.get_value().trim().is_empty() {
+        message_from_args.get_value()
+    } else {
+        message_from_json.content.as_str()
+    };
+
+    Message {
+        content: message.to_string(),
+        receivers: message_from_json.receivers.clone(),
+        fixed_hashtags: message_from_json.fixed_hashtags.clone(),
+    }
 }