+++ src/mastodon/presentation/message_resolver.rs
@@ -4,6 +4,6 @@ use crate::mastodon::application::send_message_workflow::send_message;
 
 pub fn post(args: &Args) -> Result<bool, curl::Error> {
     let config = login();
-    let result = send_message(&config, &args)?;
+    let result = send_message(&config, args)?;
     Ok(result)
 }